requirements.txt
streamlit>=1.26.0
pandas>=1.5.0
pillow>=9.0.0
gspread>=5.10.0
google-auth>=2.22.0
gspread-dataframe>=3.3.1
pytz
python-dateutil
pyyaml